+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTable.java
@@ -1190,6 +1190,7 @@ public final class FilterTable<R, C> extends JTable {
 								.model(condition)
 								.components(components)
 								.tableColumn(columnModel().column(identifier))
+								.name(identifier.toString())
 								.build());
 			}
 		}